backend/requirements.txt
# Dashboard Backend Requirements

# Web framework
flask>=3.0.0
flask-cors>=4.0.0
flask-sqlalchemy>=3.1.0
flask-socketio>=5.3.0

# Database
pymysql>=1.1.0
cryptography>=41.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
flask-migrate>=4.0.0

# Utilities
python-dotenv>=1.0.0
python-dateutil>=2.8.0

# Authentication
PyJWT>=2.8.0
Werkzeug>=3.0.0

# Production server
gunicorn>=21.0.0
eventlet>=0.33.0

# Vision Model (SmolVLM2 + Florence-2)
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.36.0
accelerate>=0.25.0
Pillow>=10.0.0
num2words>=0.5.13
einops>=0.7.0
timm>=0.9.0

# Image processing
opencv-python-headless>=4.8.0
numpy>=1.24.0
